changed to use bigint

+++ game.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, PickleType
+from sqlalchemy import Column, Integer, PickleType, BigInteger
 from sqlalchemy.orm import relationship
 
 from base import Base
@@ -7,7 +7,7 @@ from base import Base
 class Game(Base):
     __tablename__ = "games"
 
-    group_tele_id = Column(Integer, primary_key=True)
+    group_tele_id = Column(BigInteger, primary_key=True)
     game_round = Column(Integer)
     curr_player = Column(Integer)
     biggest_player = Column(Integer)
